Coursework.
The project is using core architecture with lot of code we studed during our course.
The following is chenged/added:

1. images of the products - added
2. site menu - added
3. "about" block - added
4. structure of product JSON - changed
5. logging - changed (in different folder)
6. cart box - added(as required: with additional details to the cart information)
7. product detailed information component - added as required
7. feedback form - added
8. footer - added
9. some bugs related to the asinc server requests - fixed.
10. application is fully bundled with webpack (some changes in provided configuration added)
Attention! to test bundling please check configuration file: buildDev & buildProd     commands configured for the windows environment usage - need to be fixed for MacOS, Linux or other OS. More universal configuration using clear plugin to be added in the following version.